# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Laravel API",
        "description": "API documentation for your Laravel project",
        "contact": {
            "email": "your@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions/status": {
            "get": {
                "tags": [
                    "TransactionStatus"
                ],
                "summary": "Get all transaction statuses",
                "operationId": "9e0e0fb908be47f0909c05475dbed0a4",
                "responses": {
                    "200": {
                        "description": "List of transaction statuses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransactionStatus"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TransactionStatus"
                ],
                "summary": "Reference an existing transaction status by name",
                "operationId": "54fcf80dd6636c89b22ce6cdb8c9e81e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction Status has been referenced successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/transactions/status/{id}": {
            "get": {
                "tags": [
                    "TransactionStatus"
                ],
                "summary": "Get a specific transaction status by ID",
                "operationId": "16c3a20ef25793d75015e89352c9445c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction status found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionStatus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction Status not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TransactionStatus"
                ],
                "summary": "Delete a transaction status",
                "operationId": "4ea726684371e4987feb243280747229",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction Status has been deleted"
                    },
                    "404": {
                        "description": "Transaction Status not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "TransactionStatus"
                ],
                "summary": "Update a transaction status",
                "operationId": "7339c98b625ab7e3306462bc03ad5f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "delivered"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction Status has been updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Transaction Status not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransactionStatus": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "base64",
                "in": "header"
            }
        }
    },
...
}